+++ src/components/BurgerBar/index.tsx
@@ -1,9 +1,6 @@
 import { useEffect, useState } from "react";
-import { useSession } from "next-auth/react";
 import { X } from "lucide-react";
 import BurgerItem, { BurgerbarItemKey } from "./BurgerItem";
-import { convertToAppRole } from "@/utils/utils";
-import { UserService } from "@/services/UserService";
 
 const burgerItemKeysByRole: Record<UserRole, BurgerbarItemKey[]> = {
   Attendee: ["groupcv", "groupevents", "groupcompanies", "groupgeneral"],
@@ -13,34 +10,24 @@ const burgerItemKeysByRole: Record<UserRole, BurgerbarItemKey[]> = {
 };
 
 interface BurgerBarProps {
-    onCloseAction: () => void;
+  userRole: UserRole;
+  onCloseAction: () => void;
 }
 
-export default function BurgerBar({ onCloseAction }: BurgerBarProps) {
-  const { data: session } = useSession();
+export default function BurgerBar({ userRole, onCloseAction }: BurgerBarProps) {
   const [burgerItemKeys, setBurgerItemKeys] = useState<BurgerbarItemKey[]>([]);
   const [isVisible, setIsVisible] = useState(false);
 
   useEffect(() => {
-    async function fetchUserRole() {
-      if (session?.user) {
-        const user = await UserService.getMe(session.cannonToken);
-        if (user) {
-          const userRole = convertToAppRole(user.role);
-          setBurgerItemKeys(burgerItemKeysByRole[userRole]);
-        }
-      }
-      setIsVisible(true);
-    }
-    fetchUserRole();
-  }, [session]);
+    setBurgerItemKeys(burgerItemKeysByRole[userRole]);
+    setIsVisible(true);
+  }, [userRole]);
+
 
   const handleClose = () => {
     setIsVisible(false);
     setTimeout(onCloseAction, 300);
   };
-
-  if (!session?.user) return null;
 
   return (
     <div className={`fixed inset-0 bg-[rgb(50,51,99)] z-50 flex flex-col transition-opacity duration-300 ${isVisible ? 'opacity-100 pointer-events-auto' : 'opacity-0 pointer-events-none'}`}>
@@ -49,7 +36,7 @@ export default function BurgerBar({ onCloseAction }: BurgerBarProps) {
       </div>
       <div className="flex-1 overflow-y-auto p-4">
         {burgerItemKeys.map((key) => (
-          <div key={key} className="mb-8"> {/* Aumentar o espaçamento entre os grupos */}
+          <div key={key} className="mb-8">
             <BurgerItem name={key} />
           </div>
         ))}
